% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.english}
\alias{as.english}
\alias{as.English}
\title{Numbers to text}
\usage{
as.english(x)

as.English(x)
}
\arguments{
\item{x}{integer to convert to English}
}
\description{
wrapper for english::as.english that forces character output
}
\section{Functions}{
\itemize{
\item \code{as.English}: wrapper for english::as.english that forces character output in sentence case
}}

