% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{float_tex}
\alias{float_tex}
\title{Turn floating in latex off}
\usage{
float_tex(tex, clean = TRUE)
}
\arguments{
\item{tex}{character; tex path and filename to read}

\item{clean}{boolean; clean up intermediate files?}
}
\description{
Forces latex to position figures in exact place rather than floating to "optimal" position
}
