% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insist.R
\name{insist}
\alias{insist}
\title{insist}
\usage{
insist(., test, continue = TRUE)
}
\arguments{
\item{.}{object}

\item{test}{logical test, expression}

\item{continue}{if true prints error, if false throws error}
}
\value{
the object
}
\description{
Insists a object meets a given criteria
}
\details{
Tests if the object meets the citeria. If it does, the object is returned. If not, an error is thrown unless continue is true, in which case the error is returned. Test should be an expression that refers to the object as .
}
\examples{
require(magrittr)
pi \%>\% insist(. > 3) 
pi \%>\% insist(length(.) == 1)
}
