% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_citations.R
\name{package_citations}
\alias{package_citations}
\title{package_citations}
\usage{
package_citations(packages, old_bib, new_bib)
}
\arguments{
\item{packages}{vector of packages for which citations are required}

\item{old_bib}{old bibtex file to copy}

\item{new_bib}{new bibtex file}
}
\value{
new_bib is returned invisibly
}
\description{
Append package citations to copy of main bibliography
}
